pyproject.toml
[tool.poetry]
name = "fastapi-sample"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
uvicorn = "^0.29.0"
pydantic = "^2.7.1"
pydantic-settings = "^2.2.1"
alembic = "^1.13.1"
aiomysql = "^0.2.0"
sqlalchemy = "^2.0.29"
cryptography = "^42.0.5"


[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.23.6"
httpx = "^0.27.0"
pytest-mock = "^3.14.0"

[tool.pytest.ini_options]
pythonpath = "src"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
